+++ databases/Lines/avtLinesFileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2016, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -293,7 +293,7 @@ avtLinesFileFormat::ReadFile(void)
                 lineNames.push_back(headerName);
                 headerName = "";
             }
-            int len = xl.size();
+            size_t len = xl.size();
             bool shouldAddPoint = true;
             if (len > 0)
             {
