+++ databases/MFEM/JSONRoot.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2016, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -568,7 +568,7 @@ JSONRoot::DataSet(const std::string &dset_name)
 int
 JSONRoot::NumberOfDataSets() const
 {
-    return dsets.size();
+    return (int)dsets.size();
 }
