+++ databases/Point3D/avtPoint3DFileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2016, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -419,7 +419,7 @@ avtPoint3DFileFormat::ReadData(void)
             
     }
 
-    int npts = var1.size();
+    int npts = (int)var1.size();
 
     column1 = vtkFloatArray::New();
     column1->SetNumberOfTuples(npts);
